+++ RELEASE-NOTES.md
@@ -44,6 +44,7 @@
   * Introduced new `deploy` command for project with procedures and functions.
   * Introduced new `build` command for project with procedure and functions
 * Added support for external access integration for functions and procedures
+* Added support for runtime version in snowpark procedures ad functions.
 
 ## Fixes and improvements
 * Allow the use of quoted identifiers in stages